renamed files, added swipable login page and homepage
