+++ mobile-app/lib/src/app.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:golocal/src/auth/auth_repository.dart';
+import 'package:golocal/src/auth/ui/auth_screen.dart';
 import 'package:golocal/src/event/bloc/events_bloc.dart';
 import 'package:golocal/src/event/data/events_repository_dummy.dart';
 import 'package:golocal/src/event/ui/events_view_page.dart';
 import 'package:golocal/src/auth/bloc/auth_bloc.dart';
-import 'package:golocal/src/auth/ui/signin_page.dart';
+import 'package:golocal/src/auth/ui/signin_form.dart';
+import 'package:golocal/src/home/home_page.dart';
 
 class GoLocalApp extends StatelessWidget {
   const GoLocalApp({super.key});
@@ -27,14 +29,11 @@ class GoLocalApp extends StatelessWidget {
         child: MaterialApp(
           debugShowCheckedModeBanner: false,
           home: BlocConsumer<AuthBloc, AuthState>(
+            listenWhen: (previous, current) => current is AuthError,
             listener: (context, state) {
               if (state is AuthError) {
                 ScaffoldMessenger.of(context)
                     .showSnackBar(SnackBar(content: Text(state.message)));
-              }
-              if (state is Authenticated) {
-                BlocProvider.of<EventsBloc>(context)
-                    .add(const FetchEvents(refresh: true));
               }
             },
             builder: (context, state) {
@@ -42,9 +41,11 @@ class GoLocalApp extends StatelessWidget {
                 return const SplashScreen();
               }
               if (state is Authenticated) {
-                return EventsViewPage();
+                BlocProvider.of<EventsBloc>(context)
+                    .add(const FetchEvents(refresh: true));
+                return HomePage();
               }
-              return SignInPage();
+              return AuthScreen();
             },
           ),
         ),
